requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
dm-tree==0.1.8
flatbuffers==24.3.25
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
idna==3.7
kagglehub==0.2.5
keras==2.15.0
keras-core==0.1.7
keras-nlp==0.10.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-text==2.15.0
termcolor==2.4.0
tf_keras==2.15.1
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.14.1
